+++ ddtrace/contrib/pytest/_plugin_v2.py
@@ -522,7 +522,7 @@ def _pytest_terminal_summary_post_yield(terminalreporter, failed_reports_initial
 @pytest.hookimpl(hookwrapper=True, tryfirst=True)
 def pytest_terminal_summary(terminalreporter, exitstatus, config):
     """Report flaky or failed tests"""
-    from ddtrace.contrib.pytest.plugin import print_iast_report
+    from ddtrace.appsec._iast._pytest_plugin import print_iast_report
 
     print_iast_report(terminalreporter)
