+++ ddtrace/contrib/pytest/plugin.py
@@ -20,6 +20,7 @@ from ddtrace.appsec._iast._utils import _is_iast_enabled
 from ddtrace.contrib.pytest._utils import _USE_PLUGIN_V2
 from ddtrace.contrib.pytest._utils import _extract_span
 from ddtrace.contrib.pytest._utils import _pytest_version_supports_itr
+from ddtrace.appsec._iast._pytest_plugin import ddtrace_iast
 
 
 DDTRACE_HELP_MSG = "Enable tracing of pytest functions."
@@ -140,96 +141,6 @@ def ddspan(request):
 
     if _CIVisibility.enabled:
         return _extract_span(request.node)
-
-
-vuln_data = []
-host = "docs.datadoghq.com"
-
-remediation = {
-    "SQL_INJECTION": f"https://{host}/vulnerabilities/sql_injection/sql_injection_python.html#prevention",
-    "WEAK_HASH": f"https://{host}/vulnerabilities/broken_cryptography/weak_hashing_algorithm_vulnerability.html#prevention",
-}
-
-
-def extract_code_snippet(filepath, line_number, context=3):
-    """Extracts code snippet around the given line number."""
-    try:
-        with open(filepath, "r") as file:
-            lines = file.readlines()
-            start = max(0, line_number - context - 1)
-            end = min(len(lines), line_number + context)
-            return lines[start:end], start  # Return lines and starting line number
-    except Exception as e:
-        return [f"Error reading file {filepath}: {e}"], None
-
-
-def print_iast_report(terminalreporter):
-    if not _is_iast_enabled():
-        return
-
-    terminalreporter.write("\nDatadog Code Security Report:\n", bold=True, purple=True)
-
-    if vuln_data:
-        max_nodeid = max(len(entry["nodeid"]) for entry in vuln_data)
-        max_vuln_type = max(len(entry["vulnerability"]) for entry in vuln_data)
-        max_file = max(len(entry["file"]) for entry in vuln_data)
-
-        terminalreporter.write("=" * 80 + "\n")
-
-        for entry in vuln_data:
-            terminalreporter.write(f"Test: {entry['nodeid']}\n", bold=True)
-            critical = entry["vulnerability"] == "SQL_INJECTION"
-            terminalreporter.write(
-                f"Vulnerability: {entry['vulnerability']} - \033]8;;{remediation[entry['vulnerability']]}\033\\Remediation\033]8;;\033\\ \n",
-                bold=True,
-                red=critical,
-                yellow=not critical,
-            )
-            terminalreporter.write(f"Location: {entry['file']}:{entry['line']}\n")
-            terminalreporter.write("Code:\n")
-            code_snippet, start_line = extract_code_snippet(entry["file"], entry["line"])
-            if start_line is not None:
-                for i, line in enumerate(code_snippet, start=start_line + 1):
-                    if i == entry["line"]:
-                        terminalreporter.write(f"{i:4d}: {line}", bold=True, purple=True)
-                    else:
-                        terminalreporter.write(f"{i:4d}: {line}")
-            else:
-                # If there's an error extracting the code snippet
-                terminalreporter.write(code_snippet[0] + "\n", bold=True)
-
-            # terminalreporter.write(f"\n")
-            terminalreporter.write("=" * 80 + "\n")
-
-    else:
-        terminalreporter.write("\nNo vulnerabilities found.\n")
-
-
-@pytest.fixture(autouse=_is_iast_enabled())
-def ddtrace_iast(request, ddspan):
-    """Return the :class:`ddtrace._trace.span.Span` instance associated with the
-    current test when Datadog CI Visibility is enabled.
-    """
-    yield
-    data = ddspan.get_tag(IAST.JSON)
-    if data:
-        import json
-
-        json_data = json.loads(data)
-
-        if json_data["vulnerabilities"]:
-            for vuln in json_data["vulnerabilities"]:
-                vuln_data.append(
-                    {
-                        "nodeid": request.node.nodeid,
-                        "vulnerability": vuln["type"],
-                        "file": vuln["location"]["path"],
-                        "line": vuln["location"]["line"],
-                    }
-                )
-            if request.config.getoption("ddtrace-iast-fail-tests"):
-                vulns = ", ".join([vuln["type"] for vuln in json_data["vulnerabilities"]])
-                pytest.fail(f"There are vulnerabilities in the code: {vulns}")
 
 
 @pytest.fixture(scope="session")
